+++ src/pageEditor/slices/editorSlice.ts
@@ -79,7 +79,6 @@ import { getCurrentURL, thisTab } from "@/pageEditor/utils";
 import { resolveExtensionInnerDefinitions } from "@/registry/internal";
 import { QuickBarStarterBrickABC } from "@/starterBricks/quickBarExtension";
 import { testMatchPatterns } from "@/bricks/available";
-import { type BaseExtensionPointState } from "@/pageEditor/starterBricks/elementConfig";
 import { BusinessError } from "@/errors/businessErrors";
 import { serializeError } from "serialize-error";
 import { isModComponentBase } from "@/pageEditor/sidebar/common";
@@ -93,6 +92,9 @@ import { type ModComponentBase } from "@/types/modComponentTypes";
 import { type OptionsArgs } from "@/types/runtimeTypes";
 import { createMigrate } from "redux-persist";
 import { migrations } from "@/store/editorMigrations";
+import { StarterBrick } from "@/types/starterBrickTypes";
+import { getErrorMessage } from "@/errors/errorHelpers";
+import { BaseExtensionPointState } from "@/pageEditor/baseFormStateTypes";
 
 export const initialState: EditorState = {
   selectionSeq: 0,
@@ -165,7 +167,20 @@ const checkAvailableInstalledExtensions = createAsyncThunk<
 >("editor/checkAvailableInstalledExtensions", async (arg, thunkAPI) => {
   const elements = selectNotDeletedElements(thunkAPI.getState());
   const extensions = selectNotDeletedExtensions(thunkAPI.getState());
-  const extensionPoints = await getInstalledExtensionPoints(thisTab);
+  console.debug(
+    "checkAvailableInstalledExtensions - getting installed extension points"
+  );
+  let extensionPoints: StarterBrick[];
+  try {
+    extensionPoints = await getInstalledExtensionPoints(thisTab);
+  } catch (error) {
+    console.error(getErrorMessage(error));
+    throw error;
+  }
+
+  console.debug(
+    "checkAvailableInstalledExtensions - finished getting extension points"
+  );
   const installedExtensionPoints = new Map(
     extensionPoints.map((extensionPoint) => [extensionPoint.id, extensionPoint])
   );
@@ -980,5 +995,5 @@ export const persistEditorConfig = {
   storage: localStorage as StorageInterface,
   version: 2,
   migrate: createMigrate(migrations, { debug: Boolean(process.env.DEBUG) }),
-  blacklist: ["isVarPopoverVisible"],
+  blacklist: ["isVarPopoverVisible", "error"],
 };
