+++ src/pageEditor/starterBricks/quickBar.ts
@@ -33,10 +33,7 @@ import {
 import { omitEditorMetadata } from "./pipelineMapping";
 import { type StarterBrickConfig } from "@/starterBricks/types";
 import { faThLarge } from "@fortawesome/free-solid-svg-icons";
-import {
-  type ElementConfig,
-  type SingleLayerReaderConfig,
-} from "@/pageEditor/starterBricks/elementConfig";
+import { type ElementConfig } from "@/pageEditor/starterBricks/elementConfig";
 import {
   type QuickBarConfig,
   type QuickBarDefinition,
@@ -45,6 +42,7 @@ import {
 import QuickBarConfiguration from "@/pageEditor/tabs/quickBar/QuickBarConfiguration";
 import type { DynamicDefinition } from "@/contentScript/pageEditor/types";
 import { type QuickBarFormState } from "./formStateTypes";
+import { SingleLayerReaderConfig } from "@/pageEditor/baseFormStateTypes";
 
 function fromNativeElement(url: string, metadata: Metadata): QuickBarFormState {
   const base = makeInitialBaseState();
