+++ src/pageEditor/tabs/effect/BlockPreview.tsx
@@ -40,7 +40,6 @@ import { createSlice, type PayloadAction } from "@reduxjs/toolkit";
 import { makeServiceContext } from "@/services/integrationUtils";
 import getType from "@/runtime/getType";
 import { type BrickType } from "@/runtime/runtimeTypes";
-import { type BaseExtensionPointState } from "@/pageEditor/starterBricks/elementConfig";
 import { DataPanelTabKey } from "@/pageEditor/tabs/editTab/dataPanel/dataPanelTypes";
 import DataTabJsonTree from "@/pageEditor/tabs/editTab/dataPanel/DataTabJsonTree";
 import { type RegistryId } from "@/types/registryTypes";
@@ -48,6 +47,7 @@ import { type Brick } from "@/types/brickTypes";
 import { type ApiVersion, type BrickArgsContext } from "@/types/runtimeTypes";
 import { type IntegrationDependency } from "@/types/integrationTypes";
 import { type ModComponentFormState } from "@/pageEditor/starterBricks/formStateTypes";
+import { BaseExtensionPointState } from "@/pageEditor/baseFormStateTypes";
 
 /**
  * Bricks to preview even if there's no trace.
