+++ src/store/editorMigrations.test.ts
@@ -22,10 +22,6 @@ import {
 } from "@/store/editorMigrations";
 import { initialState } from "@/pageEditor/slices/editorSlice";
 import { mapValues, omit } from "lodash";
-import {
-  type BaseFormStateV1,
-  type BaseFormStateV2,
-} from "@/pageEditor/starterBricks/elementConfig";
 import { formStateFactory } from "@/testUtils/factories/pageEditorFactories";
 import {
   type IntegrationDependencyV1,
@@ -35,6 +31,10 @@ import { integrationDependencyFactory } from "@/testUtils/factories/integrationF
 import { uuidSequence } from "@/testUtils/factories/stringFactories";
 import { modMetadataFactory } from "@/testUtils/factories/modComponentFactories";
 import { validateRegistryId } from "@/types/helpers";
+import {
+  BaseFormStateV1,
+  BaseFormStateV2,
+} from "@/pageEditor/baseFormStateTypes";
 
 const initialStateV1: PersistedEditorStateV1 = {
   ...omit(initialState, "elements", "deletedElementsByRecipeId"),
