+++ src/store/extensionsSlice.ts
@@ -28,7 +28,6 @@ import { uuidv4 } from "@/types/helpers";
 import { cloneDeep, partition } from "lodash";
 import { saveUserExtension } from "@/services/apiClient";
 import reportError from "@/telemetry/reportError";
-import { type ModComponentState } from "@/store/extensionsTypes";
 import { type Except } from "type-fest";
 import { assertModComponentNotResolved } from "@/runtime/runtimeUtils";
 import { revertAll } from "@/store/commonActions";
@@ -49,10 +48,7 @@ import {
   getIntegrationIds,
   pickModDefinitionMetadata,
 } from "@/utils/modDefinitionUtils";
-
-export const initialState: ModComponentState = {
-  extensions: [],
-};
+import { initialState } from "@/store/extensionsSliceInitialState";
 
 type ActivateModComponentParam = {
   modComponentDefinition: ModComponentDefinition;
